[CHANGED]: - Code Netbeans 8.0 Java 7 Refactoring. - Put "Working" Label to "Load Media Folder" operation (Thanks Apollo FM)
[ADDED]:
- SourceDesktop Capture custom rate setting (Thanks William for the precious diff Patches) 
- StreamPanelText Local-Time-Display (Thanks Angelocracy Xue)
- StramPanelQRCode Live-Time-Display (No Streaming Software has That :D ...)
- "Lock Mixer Aspect Ratio" button in MasterPanel.
- Moving Channels Up/Down (Works only in Sequential Playlists)
- "Apply Selected Transition to All Channels" button (Thanks Apollo FM)

[FIXED]:
- ProcessRenderer HotFix. (Missed in the last build ...)
- "Apply to Stream" MasterPanel Button doesn't works with SourceText or SourceQRCode.
- Check for duplicates names adding a channel.
- "Automatic Playlist" Operation records wrong channels for the loaded stream using "Load Media  Folder".

[TODO]:
- "Playlist Editor for WS" (Thanks Apollo FM)
- fix webcam count if error writing occurs ...
- fix radioacTV effect at 720x480 ... (It's a Gstreamer issue)
- BackEnds AutoSwitch based on distro's default backends availability.
- Technically, ProcessRenderer should be modified to launch each existing
process and then Capturer should read from the InputStream of the process
instead of reading from a local TCPIP port. (Patrick Balleux)
- Adding Audio Input/Output selector (or something similar).

+++ src/com/jhlabs/image/PerspectiveFilter.java
@@ -122,7 +122,7 @@ public class PerspectiveFilter extends TransformFilter {
      * @return the X origin.
      */
 	public float getOriginX() {
-		return x0 - (int)Math.min( Math.min( x0, x1 ), Math.min( x2, x3 ) );
+		return x0 - Math.min( Math.min( x0, x1 ), Math.min( x2, x3 ) );
 	}
 
     /**
@@ -130,7 +130,7 @@ public class PerspectiveFilter extends TransformFilter {
      * @return the Y origin.
      */
 	public float getOriginY() {
-		return y0 - (int)Math.min( Math.min( y0, y1 ), Math.min( y2, y3 ) );
+		return y0 - Math.min( Math.min( y0, y1 ), Math.min( y2, y3 ) );
 	}
 
 /*
